refactor(resolver): consolidate file specifier normalization
Fold the Windows drive-letter casing fix and the redundant-segment
collapse into a single normalize_resolved_file_specifier helper in
workspace.rs, applied at the DenoResolver::resolve chokepoint that every
resolution branch (workspace, node, npm) funnels through. This extends
the segment-collapse normalization to the node/npm resolution branches,
which previously only ran it on the workspace branch.
